requirements.txt

torch
torchvision
torchxrayvision
scikit-image
numpy
apache-airflow
apache-airflow-providers-docker
apache-airflow-providers-amazon
pyspark
boto3
